Updated tmdisplay.js
Massive update. Away from a computer with git, hence the lack of commits. This adds labels to the nodes and an inner circle so that reject and accept states can be created just by classing the group correctly. Added the start of the code needed for start states. The code finally handles update from enter update exit properly.